backend/requirements.txt
fastapi==0.105.0
uvicorn
pymongo
python-dotenv
openai
asyncio